+++ virtual-slave/src/main/java/br/com/slave/configuration/SlaveConfiguration.java
@@ -84,8 +84,8 @@ public class SlaveConfiguration extends CommonConfiguration {
     private void volumePopulator(IVolume volumeBusiness, Environment env) {
         VolumeTO volume = new VolumeTO();
         volume.setLocalizacao(env.getRequiredProperty(VOLUME_LOCALIZACAO));
-        volume.setCapacidade(Long.valueOf(env.getRequiredProperty(VOLUME_CAPACIDADE)));
-        volume.setTamanho(0L);
+        volume.setCapacidade(Integer.valueOf(env.getRequiredProperty(VOLUME_CAPACIDADE)));
+        volume.setTamanho(0);
         volume.setContem(0);
         volume.setDisponibilidade(true);
         volumeBusiness.incluir(volume);
